Added rolling view to performance analysis and refactored some code

+++ pkg/core/messaging/inventory/model.go
@@ -48,9 +48,9 @@ type BrokerProduct struct {
 	Username         string   `json:"username"`
 	Symbol           string   `json:"symbol"`
 	Name             string   `json:"name"`
-	PointValue       float32  `json:"pointValue"`
-	CostPerOperation float32  `json:"costPerOperation"`
-	MarginValue      float32  `json:"marginValue"`
+	PointValue       float64  `json:"pointValue"`
+	CostPerOperation float64  `json:"costPerOperation"`
+	MarginValue      float64  `json:"marginValue"`
 	Increment        float64  `json:"increment"`
 	MarketType       string   `json:"marketType"`
 	ProductType      string   `json:"productType"`
